insight_loop/requirements.txt
# Core ADK dependencies
google-adk>=0.3.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# SQL engine
duckdb>=0.9.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Development tools
pytest>=7.4.0
pytest-asyncio>=0.21.0
